Organized, CMake added
