+++ src/cameraLightColor/camera.cpp
