+++ src/cameraLightColor/light.cpp
