+++ src/cameraLightColor/light.h
@@ -1,7 +1,7 @@
 #ifndef LIGHT
 #define LIGHT
 
-#include "vec3d.h"
+#include "utils/vec3d.h"
 
 class Light {
 public:
