+++ src/collections/collection_lights.cpp
