+++ src/collections/collection_textures.cpp
