+++ src/utils/PathTracer.hpp
@@ -1,0 +1,6 @@
+#include "path_tracer_constants.h"
+#include "path_tracer.h"
+#include "collections/collection_hittables.h"
+#include "collections/collection_materials.h"
+#include "collections/collection_lights.h"
+#include "collections/collection_instances.h"
