+++ src/utils/ray.cpp
